Restore logging coverage, return core email messages to email service with coverage there.
